CON-646: preventing migration erros with no users data

+++ database/migrations/2020_09_01_200031_populate_jobs_outbox.php
@@ -17,20 +17,21 @@ class PopulateJobsOutbox extends Migration
     public function up()
     {
         $user = User::first();
-        $s3Configuration = S3Configuration::create([
-            'url' => 'https://s3.osl1.safedc.net',
-            'key_id' => 'A027DQI8VXPIJETYNXZQ',
-            'secret' => 'OOE4owN4uin0ctQQ6VAqsDmsHnGh4AUjgrsbEFtg',
-            'bucket' => 'jobs-outbox'
-        ]);
-
-        $storageLocation = StorageLocation::create([
-            'locatable_type' => 'App\S3Configuration',
-            'locatable_id' => $s3Configuration->id,
-            'owner_id' => $user->id,
-            'storable_type' => 'App\Jobs',
-            'human_readable_name' => 'Jobs Outbox'
-        ]);
+        if ($user) {
+            $s3Configuration = S3Configuration::create([
+                'url' => 'https://s3.osl1.safedc.net',
+                'key_id' => 'A027DQI8VXPIJETYNXZQ',
+                'secret' => 'OOE4owN4uin0ctQQ6VAqsDmsHnGh4AUjgrsbEFtg',
+                'bucket' => 'jobs-outbox'
+            ]);
+            $storageLocation = StorageLocation::create([
+                'locatable_type' => 'App\S3Configuration',
+                'locatable_id' => $s3Configuration->id,
+                'owner_id' => $user->id,
+                'storable_type' => 'App\Jobs',
+                'human_readable_name' => 'Jobs Outbox'
+            ]);
+        }
     }
 
     /**
@@ -41,8 +42,12 @@ class PopulateJobsOutbox extends Migration
     public function down()
     {
         $storageLocation = StorageLocation::where('storable_type', 'App\Jobs')->first();
-        $storageLocation->forceDelete();
-        $s3Configuration = S3Configuration::where('bucket', 'jobs-outbox')->first();
-        $s3Configuration->forceDelete();
+        if ($storageLocation) {
+            $storageLocation->forceDelete();
+            $s3Configuration = S3Configuration::where('bucket', 'jobs-outbox')->first();
+            if ($s3Configuration) {
+                $s3Configuration->forceDelete();
+            }
+        }
     }
 }
